add immediate buy check (optional)

+++ models/helper/TelegramBotHelper.py
@@ -1,6 +1,6 @@
+from json.decoder import JSONDecodeError
 import os
 import json
-import subprocess
 from datetime import datetime
 
 from pandas.core.frame import DataFrame
@@ -17,48 +17,49 @@ class TelegramBotHelper:
         self.botpath = os.path.join(self.app.telegramdatafolder, self.botfolder, self.market)
         self.filename = self.market + ".json"
 
-        if not os.path.exists(self.botfolder):
-            os.makedirs(self.botfolder)
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl():
+            if not os.path.exists(self.botfolder):
+                os.makedirs(self.botfolder)
 
-        self.data = {}
+            self.data = {}
 
-        if not os.path.exists(os.path.join(self.app.telegramdatafolder, "telegram_data")):
-            os.mkdir(os.path.join(self.app.telegramdatafolder, "telegram_data"))
+            if not os.path.exists(os.path.join(self.app.telegramdatafolder, "telegram_data")):
+                os.mkdir(os.path.join(self.app.telegramdatafolder, "telegram_data"))
 
-        if os.path.isfile(os.path.join(self.app.telegramdatafolder, "telegram_data", self.filename)):
-            self._read_data()
-        else:
-            ds = {'botcontrol' : {"status":"active", "manualsell" : False, "manualbuy" : False, "started": datetime.now().isoformat()}}
-            self.data = ds
-            self._write_data()
+            if os.path.isfile(os.path.join(self.app.telegramdatafolder, "telegram_data", self.filename)):
+                self._read_data()
+            else:
+                ds = {'botcontrol' : {"status":"active", "manualsell" : False, "manualbuy" : False, "started": datetime.now().isoformat()}}
+                self.data = ds
+                self._write_data()
 
-        if os.path.isfile(os.path.join(self.app.telegramdatafolder, "telegram_data", "data.json")):
-            self._read_data("data.json")
-            if not "markets" in self.data:
-                self.data.update({"markets": {}})
+            if os.path.isfile(os.path.join(self.app.telegramdatafolder, "telegram_data", "data.json")):
+                self._read_data("data.json")
+                if not "markets" in self.data:
+                    self.data.update({"markets": {}})
+                    self._write_data("data.json")
+            else:
+                ds = {"trades" : {}, "markets": {}}
+                self.data = ds
                 self._write_data("data.json")
-        else:
-            ds = {"trades" : {}, "markets": {}}
-            self.data = ds
-            self._write_data("data.json")
 
     def _read_data(self, name: str = "") -> None:
         file = self.filename if name =="" else name
-        with open(os.path.join(self.app.telegramdatafolder, 'telegram_data', file), 'r') as json_file:
+        with open(os.path.join(self.app.telegramdatafolder, 'telegram_data', file), 'r', encoding='utf8') as json_file:
             self.data = json.load(json_file)
 
     def _write_data(self, name: str = "") -> None:
         file = self.filename if name =="" else name
         try:
-            with open(os.path.join(self.app.telegramdatafolder, 'telegram_data', file), 'w') as outfile:
+            with open(os.path.join(self.app.telegramdatafolder, 'telegram_data', file), 'w', encoding='utf8') as outfile:
                 json.dump(self.data, outfile, indent=4)
-        except Exception as err:
+        except (JSONDecodeError, Exception) as err:
             Logger.critical(str(err))
-            with open(os.path.join(self.app.telegramdatafolder, 'telegram_data', file), 'w') as outfile:
+            with open(os.path.join(self.app.telegramdatafolder, 'telegram_data', file), 'w', encoding='utf8') as outfile:
                 json.dump(self.data, outfile, indent=4)
         
     def addmargin(self, margin: str = "", delta: str = "", price: str = ""):
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data()
 
             addmarket = {'exchange' : self.exchange, 'margin' : margin, 'delta' : delta, 'price' : price, "df_high": " ", "from_df_high": " "}
@@ -66,14 +67,14 @@ class TelegramBotHelper:
             self._write_data()
 
     def addinfo(self, message: str = "", price: str = "", df_high: str="", from_df_high: str="") -> None:
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data()
             addmarket = {"message": message, "margin": " ", "delta": " ", "price" : price, "exchange" : self.exchange, "df_high": df_high, "from_df_high": from_df_high}
             self.data.update(addmarket)
             self._write_data()
 
     def addindicators(self, indicator, state) -> None:
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data()
             if not 'indicators' in self.data:
                 self.data.update({"indicators": {}})
@@ -82,18 +83,18 @@ class TelegramBotHelper:
             self._write_data()
 
     def deletemargin(self):
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             os.remove(os.path.join(self.app.telegramdatafolder, 'telegram_data', self.filename))
 
     def closetrade(self, ts, price, margin):
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data("data.json")
             self.data['trades'].update({ts : {"pair" : self.market, "price" : price, "margin" : margin}})
             self._write_data("data.json")
 
     def checkmanualbuysell(self) -> str:
         result = "WAIT"
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data()
 
             if len(self.data['botcontrol']) > 0:
@@ -112,21 +113,21 @@ class TelegramBotHelper:
 
     def checkbotcontrolstatus(self) -> str:
         result = "active"
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data()
             result = self.data["botcontrol"]["status"]
             
         return result
 
     def updatebotstatus(self, status) -> None:
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self._read_data()
             if not self.data["botcontrol"]["status"] == status:
                 self.data["botcontrol"]["status"] = status
                 self._write_data()
 
     def removeactivebot(self) -> None:
-        if self.app.enableTelegramBotControl():
+        if not self.app.isSimulation() and self.app.enableTelegramBotControl()():
             self.deletemargin()
 
     def save_scanner_output(self, exchange, quote, output: DataFrame) -> None:
